# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Redeemer's Forge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts with filters
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [story, devotional, article, sermon, announcement, testimonial, reflection]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: tag
          in: query
          schema:
            type: ["string", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: premium
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── VIDEOS ──────────────────────────────────────────────
...
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [likes]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

  # ─── BOOKMARKS ──────────────────────────────────────────
  /posts/{postId}/bookmark:
    post:
      operationId: bookmarkPost
      tags: [bookmarks]
      summary: Bookmark/unbookmark a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bookmark toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkResult"
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── SEARCH ──────────────────────────────────────────────
...
  /analytics/posts/{id}:
    get:
      operationId: getPostAnalytics
      tags: [analytics]
      summary: Get analytics for a specific post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAnalytics"

  # ─── SUBSCRIPTIONS ──────────────────────────────────────
...
components:
  schemas:
...
    Post:
      type: object
      required: [id, title, type, slug, excerpt, content, authorName, authorAvatar, coverImage, isPremium, isFeatured, readingTimeMinutes, viewCount, likeCount, commentCount, bookmarkCount, publishedAt, createdAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        type:        { type: string }
        slug:        { type: string }
        excerpt:     { type: string }
        content:     { type: string }
        authorName:  { type: string }
        authorId:    { type: ["string", "null"] }
        authorAvatar: { type: ["string", "null"] }
        coverImage:  { type: ["string", "null"] }
        category:    { type: ["string", "null"] }
        tags:        { type: array, items: { type: string } }
        isPremium:   { type: boolean }
        isFeatured:  { type: boolean }
        isPublished: { type: boolean }
        readingTimeMinutes: { type: integer }
        viewCount:   { type: integer }
        likeCount:   { type: integer }
        commentCount: { type: integer }
        bookmarkCount: { type: integer }
        publishedAt: { type: ["string", "null"] }
        createdAt:   { type: string }
        isLiked:     { type: ["boolean", "null"] }
        isBookmarked: { type: ["boolean", "null"] }
    PostList:
      type: object
      required: [posts, total]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
    PostInput:
      type: object
      required: [title, type, content, excerpt]
      properties:
        title:       { type: string }
        type:        { type: string }
        content:     { type: string }
        excerpt:     { type: string }
        category:    { type: ["string", "null"] }
        tags:        { type: array, items: { type: string } }
        coverImage:  { type: ["string", "null"] }
        isPremium:   { type: boolean }
        isFeatured:  { type: boolean }
        isPublished: { type: boolean }
        readingTimeMinutes: { type: ["integer", "null"] }
        publishedAt: { type: ["string", "null"] }
    PostUpdate:
      type: object
      properties:
        title:       { type: ["string", "null"] }
        content:     { type: ["string", "null"] }
        excerpt:     { type: ["string", "null"] }
        category:    { type: ["string", "null"] }
        tags:        { type: array, items: { type: string } }
        coverImage:  { type: ["string", "null"] }
        isPremium:   { type: ["boolean", "null"] }
        isFeatured:  { type: ["boolean", "null"] }
        isPublished: { type: ["boolean", "null"] }
        readingTimeMinutes: { type: ["integer", "null"] }

    # ─── Video schemas ────────────────────────────────────
...
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:     { type: boolean }
        likeCount: { type: integer }
    BookmarkResult:
      type: object
      required: [bookmarked]
      properties:
        bookmarked: { type: boolean }

    # ─── User schemas ─────────────────────────────────────
    UserProfile:
      type: object
      required: [id, name, username, email, role, isPremium, joinedAt]
      properties:
        id:            { type: string }
        name:          { type: string }
        username:      { type: ["string", "null"] }
        email:         { type: string }
        bio:           { type: ["string", "null"] }
        avatar:        { type: ["string", "null"] }
        role:          { type: string }
        isPremium:     { type: boolean }
        joinedAt:      { type: string }
        postCount:     { type: integer }
        likeCount:     { type: integer }
        bookmarkCount: { type: integer }
        darkMode:      { type: boolean }
    UserProfileUpdate:
      type: object
      properties:
        name:     { type: ["string", "null"] }
        username: { type: ["string", "null"] }
        bio:      { type: ["string", "null"] }
        avatar:   { type: ["string", "null"] }
        darkMode: { type: ["boolean", "null"] }
...
    PostAnalytics:
      type: object
      required: [postId, views, likes, comments, bookmarks, shares]
      properties:
        postId:    { type: integer }
        views:     { type: integer }
        likes:     { type: integer }
        comments:  { type: integer }
        bookmarks: { type: integer }
        shares:    { type: integer }

    # ─── Subscription schemas ─────────────────────────────
...